[Core][Common][Test] Env obj exposed and single test setting
1. Exposed the environemnt object to all the test cases and
brought volds and cleands inside this class.
2. Single test execution check while creating queue for
non disruptive test execution added.
