+++ tests/lazy_parent_test.py
@@ -24,7 +24,7 @@ class LazyParentTest(metaclass=abc.ABCMeta):
     }
 
     def __init__(self, mname: str, param_obj, volume_type: str,
-                 log_path: str, log_level: str = 'I'):
+                 env_obj, log_path: str, log_level: str = 'I'):
         """
         Lazy parent will only initialize the connection.
         """
@@ -36,15 +36,16 @@ class LazyParentTest(metaclass=abc.ABCMeta):
         self.volume_type = volume_type
         self.vol_type_inf = param_obj.get_volume_types()
         self._configure(f"{mname}-{volume_type}", server_details,
-                        client_details, log_path, log_level)
+                        client_details, env_obj, log_path, log_level)
         self.server_list = param_obj.get_server_ip_list()
         self.client_list = param_obj.get_client_ip_list()
         self.brick_roots = param_obj.get_brick_roots()
 
     def _configure(self, mname: str, server_details: dict,
-                   client_details: dict, log_path: str, log_level: str):
+                   client_details: dict, env_obj, log_path: str,
+                   log_level: str):
         machine_detail = {**client_details, **server_details}
-        self.redant = RedantMixin(machine_detail)
+        self.redant = RedantMixin(machine_detail, env_obj)
         self.redant.init_logger(mname, log_path, log_level)
         self.redant.establish_connection()
         self.test_name = mname
@@ -59,6 +60,8 @@ class LazyParentTest(metaclass=abc.ABCMeta):
         which is overridden by by child test.
         """
         try:
+            self.vol_name = (f"redant-{self.volume_type}")
+            self.mountpoint = (f"/mnt/{self.volume_type}")
             self.run_test(self.redant)
         except Exception as error:
             tb = traceback.format_exc()
